- Adding some spam protection for contact forms
- Improving translation performance for Involvements
- Resolving some of the issues with Involvement Images
- Improving build & automatic update processes

+++ build.sh
@@ -19,31 +19,33 @@ mkdir build/assets/js
 
 # install uglify and uglify the JS files.
 echo $(npm install -g uglify-js)
-uglifyjs -o build/assets/js/base-defer.min.js -- assets/js/base-defer.js
-uglifyjs -o build/assets/js/meeting-defer.min.js -- assets/js/meeting-defer.js
-uglifyjs -o build/assets/js/partner-defer.min.js -- assets/js/partner-defer.js
+uglifyjs assets/js/base-defer.js -o build/assets/js/base-defer.min.js --source-map
+uglifyjs assets/js/meeting-defer.js -o build/assets/js/meeting-defer.min.js --source-map
+uglifyjs assets/js/partner-defer.js -o build/assets/js/partner-defer.min.js --source-map
 cp -r assets build
 cd ./build || exit
-zip -r ../touchpoint-wp.zip assets
 cd ..
 
 
 # compile translations
-wget -O wp-cli.phar https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+if [ ! -f wp-cli.phar ]; then
+    wget -O wp-cli.phar https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+fi
 
 cp -r ./i18n ./build/i18n
 
 php ./wp-cli.phar i18n make-json ./build/i18n
 php ./wp-cli.phar i18n make-mo ./build/i18n
+cp ./wpml-config.xml ./build/wpml-config.xml
+cp ./composer.json ./build/composer.json
+
+cp -r ./ext ./build/ext
+cp -r ./src ./build/src
+
+find . -maxdepth 1 -iname "*.php" -exec cp {} build/ \;
+find . -maxdepth 1 -iname "*.md" -exec cp {} build/ \;
+find . -maxdepth 1 -iname "*.json" -exec cp {} build/ \;
+
 cd ./build || exit
-zip -r ../touchpoint-wp.zip i18n
+find . -exec zip ../touchpoint-wp.zip {} \;
 cd ..
-
-zip -r touchpoint-wp.zip ./ext
-zip -r touchpoint-wp.zip ./src
-
-find . -maxdepth 1 -iname "*.php" -exec zip touchpoint-wp.zip {} \;
-find . -maxdepth 1 -iname "*.md" -exec zip touchpoint-wp.zip {} \;
-find . -maxdepth 1 -iname "*.json" -exec zip touchpoint-wp.zip {} \;
-
-rm -r build
